feat: add method for premium management

+++ src/main/java/at/qe/skeleton/internal/services/CreditCardService.java
@@ -22,6 +22,9 @@ public class CreditCardService {
 
   @Autowired private CreditCardRepository creditCardRepository;
 
+  @Autowired private UserxService userxService;
+
+
   public void saveCreditCard(CreditCard creditCard) throws IllegalArgumentException {
     if (!validateDate(creditCard.getExpirationDate())) {
       throw new IllegalArgumentException("Invalid expiration date.");
@@ -34,6 +37,10 @@ public class CreditCardService {
   }
 
   public void deleteCreditCard(CreditCard creditCard) {
+    if(userxService.isPremium(creditCard.getUser())) {
+      throw new IllegalArgumentException("You are a premium user. You cannot delete your credit card.");
+    }
+
     creditCardRepository.delete(creditCard);
   }
